refactor(node): Extract `Yarn2DependencyHandler` to a dedicated file
Improve the overview a bit.
